Progress in analyzing the latest results: -Analyse of new wedge Bragg measurements are ready (despite the exact positioning of the wedge) -The matrix array is now correctly mapped into an image -The 1540 frame video of the matrix array is ready! (In movie.zip are 3 different versions, contour and pixel in different speeds)

+++ EvaluationSoftware/helper_modules.py
@@ -82,7 +82,7 @@ class LineShape:
     def __init__(self, list_of_points, distance_mode=False):
         print('A shape with connection lines between the given points will be initiated! Note that for positioning of '
               'the shape the x-positions can be scaled by setting a reference point later on.')
-        self.points = np.array(list_of_points)
+        self.points = np.array(list_of_points, dtype=np.float64)
         self.distance_mode = distance_mode
         if distance_mode:
             cache = deepcopy(self.points)
